\Part 3 Total:            12.0 of 28.0\

+++ shell.cc
@@ -263,7 +263,7 @@ static void printIndent(int indent) {
 
 std::string Shell::extractNextComponent(std::string& suffix)
 {
-  int slashIndex = suffix.find('/');
+  size_t slashIndex = suffix.find('/');
 
   std::string component;
